detalhes-do-plano

+++ app/Http/Controllers/Admin/PlanController.php
@@ -37,14 +37,21 @@ class PlanController extends Controller
         //  dd($plan);
         if(!$plan)
             return redirect()->back();
-        return view('admin.pages.plans.view', ['plan'=>$plan]); 
+        return view('admin.pages.plans.show', ['plan'=>$plan]); 
     }
 
     public function destroy($url){
-        $plan = $this->repository->where('url', $url)->first(); 
+        $plan = $this->repository
+                        ->with('details')
+                        ->where('url', $url)->first(); 
     
         if(!$plan)
             return redirect()->back();
+
+        if($plan->details->count()>0){
+            return redirect()->back()->with('error', 'Não Posso detetar - Existem detalhes vinculados a este plano');
+            
+        };
         $plan->delete();
         return redirect()->route('plans.index');
     }
